tests fixed, compilation warning removed

+++ src/engine/serialdevice.cxx
@@ -153,7 +153,7 @@ bool CSerialDevice::Write(const void *data, size_t len) {
 
     // Write the message to the standard output (blocking)
     ssize_t bytesWrote = write(_fileDesc, data, len);
-    if (bytesWrote != len) {
+    if (bytesWrote != (ssize_t)len) {
         LOGERROR(1, "Error writing to serial device.\n");
         CountWriteError();
         return false;
